Rename swap_float_int() to as_int() and as_float()
This is a proposed change and not something I feel strongly about.

I was just thinking that swap_float_int() didn't read too well for
doing both operations of interpretting a float as an int AND
interpretting an int as a float.  I think it would read better if
they are separate names.

I recognize that the old names were long winded
- reinterpret_int_as_float()
- reinterpret_float_as_int()
Some shorter names I thought of:
- int_as_float(), float_as_int()
- int_as_flt(), flt_as_int()
- int2flt(), flt2int()
- *as_float(), as_int() <- this is the one in the proposed change
- as_flt(), as_int()
- int_to_float(), float_to_int()