Abstract out running tests to run_impl()
As part of this, allow for tests to receive test inputs externally.
Not all tests are implemented in this way, meaning the functionality
to give them data is actually dummy data and is generated inside.
I wanted to fix this problem in a separate commit.

This commit mostly changes the interface of TestBase and makes the
necessary changes across all tests to still compile and generate
results.