inputGen: First attempt to get it integrated
I ran into a problem: when the compiler generates duplicate symbols
in multiple object files from templates, it then at link time will
choose one of the symbols to keep and throw away all others.  This
completely sabotauges the approach I wanted to take.

Now I want to try doing a shared library for groundtruth and calling
the functions at runtime using dlload and dlsymb.

Get a proof-of-concept of shared library usage

This means that the optimized compiled objects are stored in a shared
library.  This shared library is dynamically loaded and executed through
dlopen(), dlsym() and dlclose().  We can specify options for the shared
library to prefer its own symbols over the symbols in the executable.
An that is what we have done.

Now to clean it up and make it work for all precisions.  I also want to
test if we can send C++ objects through like std::string and
std::vector.

Get the input generator to finally be working

There are some suspicious behaviors however.  When running the
DoMatrixMultSanity test, it finds many divergences, the first of which
will be very large in error and the later ones will be very small.  This
should not be repeatable since I seed not to a constant value but from a
std::random_device.

One other thing to fix is to not have the created object files appear in
the qfpc folder.

Symbolically link source files needed

This is for the inputGen executable