Get a proof-of-concept of shared library usage
This means that the optimized compiled objects are stored in a shared
library.  This shared library is dynamically loaded and executed through
dlopen(), dlsym() and dlclose().  We can specify options for the shared
library to prefer its own symbols over the symbols in the executable.
An that is what we have done.

Now to clean it up and make it work for all precisions.  I also want to
test if we can send C++ objects through like std::string and
std::vector.