Get the input generator to finally be working
There are some suspicious behaviors however.  When running the
DoMatrixMultSanity test, it finds many divergences, the first of which
will be very large in error and the later ones will be very small.  This
should not be repeatable since I seed not to a constant value but from a
std::random_device.

One other thing to fix is to not have the created object files appear in
the qfpc folder.