Replace ResultType with vector of TestResult
The old ResultType was weird to work with, a map of pair -> pair.
It is more natural to make my own custom type and have as many
fields as I want as well as convenience functions.

This change I also implemented having a way to store comparison
data and outputting it to file.