Make test string return memory efficient
This means that immediately after calling run_impl, the string gets
dumped to file and the memory is reclaimed.

Also, when doing the comparison against ground truth, the loading of the
file is done immediately before the compare() call.

This change is mostly because these returned strings may be very large
such as 300 MB or larger.  We do not want all results that may be large
in memory at the same time, especially if we do not need to.  This is a
memory scalability problem.