compare in the ground truth executable
This is instead of running the comparison separately
in the many compiled test executables.  The main
problem with the way it used to be done is that the
compilation of the test executable may influence the
compare() method.

The ground truth executable is executed now at the
very end and all of the results from all tests are
passed in all at once.  The test result files are
updated with the values of the comparison (so they
are rewritten with that column filled in).