Makefile.in: recursive targets depend on GT_TARGET
This is not because the recursive targets really depend on GT_TARGET,
but rather that we know recursive targets need to be recompiled when
GT_TARGET needs to be recompiled.  I'm piggy-backing on the dependency
graph of GT_TARGET since it is identical to each recursive target in
this case.