test-cli: add --suffix argument for compare mode
The suffix argument allows for the user to not overwrite the files to be
compared.  Instead, the comparison result will be put into the same
filename with the suffix added to the end.  By default, it is the same
behavior as before since the default suffix is the empty string.