bisect: implement searching over intel libs
This is hard-coded to search over a list of 7 static libs.
Also, this does not work, since if we include some and not others,
then there are symbols within the first static libs that are required
and not satisfied, unless you also include the other static libs.

It needs to be fixed to do all-or-nothing.  Maybe I can come up with
something later to isolate the problem better.