bisect: revert intel libs to be static libs
This means revert to absolute paths since it is possible to
link against the shared library instead of the static one,
which would cause a runtime error since there is no rpath.

Another solution could have been to add an rpath, but there
should be no need since it should be equivalent to link to
the static library or the shared library.