bisect: rebuild local gtrun for intel library problems
This is so that the bisect file search is capable of getting further
The libraries cause at least part of the variability.  By resetting
the ground-truth results and executable to have the variability
introduced by the static libraries, we can isolate further, at least
to the file bisect step.  The symbol bisect with intel is still a
little unstable because of the inline function problem.