flitHelpers: reimplement ifopen() and ofopen()
The old implementation was a cleaner syntax and compiled fine
under newer compilers.  However, under GCC 4.8, it would not
compile because that compiler would try to use the copy
assignment operator, which I think is a compiler bug.  Regardless,
we changed the implementation to work around this issue so that
the tool could support GCC 4.8.