bisect: change bisect_search to take a function with only one argument
This is more in line with delta debugging.  It makes it so that we only
pass in the list of objects we want to test.  It is really implementation
details for the problem being addressed that says we actually care about
the complement.  Also, the complement is easy enough to compute in the
test function without adding unnecessary computational complexity