Update flit_bisect.py
Yes, it can work with any list of hashable types, but it only needs to work right now with lists of strings.  I think it's easier to change the example in the docstring.  This will not be able to memoize something that takes a list of lists, or a list of dictionaries, or even a list of sets.  That's why I was specific, and why the bisect algorithm cannot memoize in general, since it would matter what is being memoized.