Generalize flags into toml file.
The flags were hardcoded into the Makefile, now the are generated
through the toml file.

This commit removes a layer of indirection in which each
flag is stored in a variable, and the compiler's flags
are lists of variables rather than lists of flags.
This indirection is useful for mutliple things, including
names, flags with spaces, and the empty flag. Therefore,
in a future commit, this level should be reproduced
at the Python level, and re-introduced. In the meantime,
this commit will compile and run fine, so long as none
of the special cases mentioned above are hit.