Generalize flag indirection
Restored the level of indirection where each flag
is stored in a variable.

From now on, DEFFLAGS is automatically included by FLiT,
as this is almost always desired and it's hard to get out
of the toml file the way the Python code works now.

To process the flags at the Python level, dictionaries are
used, which has the side effect of causing FLiT to run the
flags in whatever order the Python dictionary decides to
give it. This feels weird with optimization levels, but
all the same tests still get run.