Clean up after messy rebase
Recently rebased this branch onto the latest
devel. There were a few conflicts that could
not be tested until the rebase was done, so
those debugs are in this commit.

They all relate to the introduction of compiler
collections in this branch, and how that conflicts
with the addition of passing a 'type' to the makefile
which as been added in devel.

As a stop-gap, the type that is used is the type of the
c++ compiler, tacitly assuming that all types in a
compiler collection are the same. Adding in
per-compiler types is a fix that should happen sooner
than later.