Add per-compiler attributes to the Makefile
Fixes #125

Also fixes the stop-gap in the last commit

Generates assingments for per-compiler attributes to
the Makefile. Right now, there are three of these:
 -- compiler type
 -- required flags
 -- required ld flags

The "type" attribute was used before, it was essentially
hard-coded in. This generalizes that behavior and allows
for types to be defined per compiler.

The other two are used to implement #125, which is compiler-
specific ld flags and requried flags.

This commit adds another level of indirection in all of the
make variables. In order to simplify things, I've tried to
use existing shortcuts and make a few new ones. Most notably,
see R_CUR_CXX and R_CUR_FXX.