flit_capture.py: add --add-{cc,cxx} and --add-lang
The --use-cc was a bit hackish.  You could only add one
thing, and that compiler was used internally to compile
the shared library to be used.  The --use-cxx was about
equally as useless.  Instead, I changed it to be a comma-
separated list of compilers you can add.

Now, you can specify a new language that is not baked into the script.
This is one of the necessary features we had for this tool.