flit_bisect: add 'auto' option for --compiler-type
The auto will try to resolve using the 'flit-config.toml' file, and if
it cannot, then it will fall back to 'misc'.

The reason for this instead of just using 'misc' is that now specifying
'misc' explicitly will result in NOT trying to resolve against
'flit-config.toml'.  This may be useful in situations where the 'smart'
decisions in flit may be doing something wrong for your use case.  You
can use this option to disable any flit 'smarts'.