Remove dependency on source files ending in .cpp
Instead use the list of files in SOURCE within the Makefiles.
We now create a rule for each file rather than relying on GNU
Makefile expanding implicit rules (such as %.o: %.cpp ...).

Fixed tests and bisect to work with this new framework.

Did not update documentation or add new tests