flit-autoconfig: allow make variables to be missing from outfile
This is when appending.  Before this change, it was required that
in append mode, the Makefile variables 'SOURCE', 'CC_REQUIRED',
and 'LD_REQUIRED' were all present in the output file (to check
for and remove duplicates).  Now, it is not an error if those
variables are not specified, as it will default to treating them
as empty.