Re-enable SimpleCHull litmus test
We fixed the problem of the external file by including the values
into the test itself.  It turns out the external file was simply
a list of 1,000 32-bit float values in binary format.  I read
them, put them in as a vector, and then write them to a temporary
file from the test to then be read by the convex hull framework.